% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litellm.R
\name{litellm_models}
\alias{litellm_models}
\title{Retrieve available models from the LiteLLM server}
\usage{
litellm_models()
}
\value{
A data frame containing information about the models available on
the LiteLLM server.
}
\description{
This function requests a list of available language models from the
LiteLLM server. It can be used to check which models are accessible
before calling functions such as \code{litellm_chat()}.
}
\details{
The function sends a request to the \verb{/models} endpoint of the
LiteLLM API and returns the response as an R list.

A connection must first be configured using \code{litellm_setup()}.
}
\examples{
\dontrun{

litellm_setup(
  api_key = "MY_API_KEY",
  base_url = "https://my-server/v1"
)

models <- litellm_models()

print(models)

}

}
